requirements.txt
numpy
matplotlib
pillow
joblib
tqdm
